Se commitean cambios, se agrega la parte de cuentas y la parte de impresión de las cuentas

+++ backend/Mova/src/main/java/com/movauy/mova/service/sale/SaleService.java
@@ -54,7 +54,7 @@ public class SaleService {
         User currentUser = authService.getUserById(userId);
 
         Sale sale = new Sale();
-        sale.setTotalAmount(saleDTO.getTotalAmount());
+        sale.setTotalAmount(saleDTO.getTotalAmount().doubleValue());
         sale.setPaymentMethod(saleDTO.getPaymentMethod());
         sale.setDateTime(LocalDateTime.now());
         sale.setCashRegister(cashReg);
